fix(nextjs,backend): Move constants in @clerk/backend
